benchmarks: merge TTS speed and WER eval scripts into unified modules
- Add benchmarks/tasks/tts.py: unified task module replacing tasks/tts_speed.py
  and tasks/voice_clone.py (ASR/WER helpers, VoiceCloneTTS, VoiceCloneOmni,
  TTS HTTP send layer, speed result builders)

- Add benchmarks/eval/benchmark_omni_seedtts.py: two-phase SeedTTS pipeline for
  Qwen3-Omni replacing benchmark_omni_tts_speed.py + voice_clone_omni_wer.py;
  generate phase always saves audio, transcribe phase reads it without the server

- Add benchmarks/eval/benchmark_tts_seedtts.py: two-phase SeedTTS pipeline for
  S2-Pro replacing benchmark_tts_speed.py + voice_clone_tts_wer.py; supports
  streaming, concurrency, --generate-only / --transcribe-only flags

- Update benchmark_omni_mmmu.py: import from benchmarks.tasks.tts instead of
  benchmarks.tasks.tts_speed and benchmarks.tasks.voice_clone

- Update test_s2pro_tts_ci.py: import TtsSeedttsBenchmarkConfig /
  run_tts_seedtts_benchmark; point WER_SCRIPT to benchmark_tts_seedtts.py

- Update test_qwen3_omni_tts_ci.py: import OmniSeedttsBenchmarkConfig /
  run_omni_seedtts_benchmark; remove separate _run_wer_generate step; reuse
  speed-benchmark audio via module-level _SPEED_OUTPUT_DIR (matching S2-Pro
  pattern) to avoid double audio generation and GPU OOM in CI

- Delete 6 superseded files: tasks/tts_speed.py, tasks/voice_clone.py,
  eval/benchmark_omni_tts_speed.py, eval/voice_clone_omni_wer.py,
  eval/benchmark_tts_speed.py, eval/voice_clone_tts_wer.py

Closes #277
