+++ benchmarks/eval/benchmark_omni_mmmu.py
@@ -37,18 +37,18 @@ from benchmarks.benchmarker.runner import BenchmarkRunner, RunConfig
 from benchmarks.benchmarker.utils import save_json_results, wait_for_service
 from benchmarks.dataset.mmmu import load_mmmu_samples
 from benchmarks.metrics.performance import compute_speed_metrics
-from benchmarks.tasks.tts_speed import print_speed_summary
-from benchmarks.tasks.visual_understand import (
-    compute_mmmu_metrics,
-    make_mmmu_send_fn,
-    print_mmmu_accuracy_summary,
-)
-from benchmarks.tasks.voice_clone import (
+from benchmarks.tasks.tts import (
     SampleOutput,
     _transcribe_and_compute_wer,
     calculate_wer_metrics,
     load_asr_model,
+    print_speed_summary,
     print_wer_summary,
+)
+from benchmarks.tasks.visual_understand import (
+    compute_mmmu_metrics,
+    make_mmmu_send_fn,
+    print_mmmu_accuracy_summary,
 )
 
 logging.basicConfig(
