+++ tests/test_model/test_s2pro_tts_ci.py
@@ -35,9 +35,9 @@ from typing import Literal
 import pytest
 
 from benchmarks.dataset.prepare import DATASETS, download_dataset
-from benchmarks.eval.benchmark_tts_speed import (
-    TtsSpeedBenchmarkConfig,
-    run_tts_speed_benchmark,
+from benchmarks.eval.benchmark_tts_seedtts import (
+    TtsSeedttsBenchmarkConfig,
+    run_tts_seedtts_benchmark,
 )
 from sglang_omni.utils import find_available_port
 from tests.test_model.conftest import (
@@ -176,7 +176,7 @@ WER_SCRIPT = str(
     Path(__file__).resolve().parents[2]
     / "benchmarks"
     / "eval"
-    / "voice_clone_tts_wer.py"
+    / "benchmark_tts_seedtts.py"
 )
 
 
@@ -198,17 +198,16 @@ def _run_benchmark(
     max_samples: int | None = None,
     stream: bool = False,
 ) -> dict:
-    benchmark_config = TtsSpeedBenchmarkConfig(
+    benchmark_config = TtsSeedttsBenchmarkConfig(
         model=S2PRO_MODEL_PATH,
         port=port,
-        testset=testset,
+        meta=testset,
         output_dir=output_dir,
         concurrency=concurrency,
         max_samples=max_samples,
-        save_audio=True,
         stream=stream,
     )
-    speed_results = asyncio.run(run_tts_speed_benchmark(benchmark_config))
+    speed_results = asyncio.run(run_tts_seedtts_benchmark(benchmark_config))
     _validate_speed_results_keys(speed_results)
     return speed_results
